Messagess

+++ src/main/java/projekt/nieruchomosci/service/BusinessService.java
@@ -8,5 +8,5 @@ public interface BusinessService {
     void add(Business business);
     void delete(Business business);
     List<Business> getAll();
-    Business findById(int id);
+    Business findById(Long id);
 }
